requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
speedtest-cli==2.1.3
urllib3==2.2.1
